Matrix 2: In MainWindowButtonDescription, declared and defined buildForShowGlobalParamsLayer(). In GUI_Layer_MainWindowButtons, added instances of GUI_Layer_GlobalParameters and GUI_Layer_CommError, added TextButton btn_ShowGlobalParams, and declared and defined prepareToShowGlobalParamsLayer(), showCommErrorLayer(), and showGlobalParamsLayer().

+++ Matrix-6G/Matrix-6G 2.0/Source/midi/midi_0_RawDataTools.cpp
@@ -421,7 +421,7 @@ const std::vector<uint8> RawDataTools::extractRawGlobalDataFrom_GUI(GlobalOption
     globalData.push_back(0);
     checksum += spilloverIsEnabled ? 1 : 0;
 
-    auto undefinedByte{ 0 };
+    auto undefinedByte{ (uint8)0 };
     globalData.push_back(undefinedByte);
     globalData.push_back(undefinedByte);
